{{ cookiecutter.repo_name }}/requirements.txt
sphinx-rtd-theme
python-dotenv
psycopg2
SQLAlchemy
